+++ stack_init.c
@@ -46,7 +46,7 @@ static void	append_node(t_stack_node **s, int n)
 
 static long	ft_atol(const char *nptr)
 {
-	long	nbr;
+	long long 	nbr;
 	int		i;
 	int		s;
 
@@ -80,7 +80,7 @@ t_stack_node	*get_cheapest(t_stack_node *stack)
 
 int	init_stack_a(t_stack_node **a, char **av)
 {
-	long			nbr;
+	long 		nbr;
 	int				i;
 
 	i = 0;
